fix: добавляет сравнение в mysqli_num_rows и упрощает вложенность

+++ subscribe.php
@@ -18,20 +18,19 @@
     $sql_subscriber = 'SELECT subscribe_id FROM subscriptions WHERE subscribe_id = ? AND follower_id = ?';
     $result = form_sql_request($con, $sql_subscriber, [$user_id, $_SESSION['user']['id']]);
 
-    if (mysqli_num_rows($result)) {
+    if (mysqli_num_rows($result) > 0) {
         $is_subscribe = true;
     }
 
-    if ($user_id) {
-        if (!$is_subscribe) {
-            $sql_subscribe = 'INSERT INTO subscriptions(subscribe_id, follower_id) VALUES(?, ' . $_SESSION['user']['id'] . ')';
-            $result = form_sql_request($con, $sql_subscribe, [$user_id], false);
-            header("Location: profile.php?user=" . $user_id . "&tab=posts");
-            exit();
-        }
+    if ($user_id && !$is_subscribe) {
+        $sql_subscribe = 'INSERT INTO subscriptions(subscribe_id, follower_id) VALUES(?, ' . $_SESSION['user']['id'] . ')';
+        $result = form_sql_request($con, $sql_subscribe, [$user_id], false);
+        header("Location: profile.php?user=" . $user_id . "&tab=posts");
+    }
 
+    if($user_id && $is_subscribe) {
         $sql_subscribe = 'DELETE FROM subscriptions WHERE subscribe_id = ? AND follower_id = ?';
-        $result = form_sql_request($con, $sql_subscribe, [$user_id, $_SESSION['user']['id']]);
+        $result = form_sql_request($con, $sql_subscribe, [$user_id, $_SESSION['user']['id']], false);
         header("Location: profile.php?user=" . $user_id . "&tab=posts");
     }
